backend/requirements.txt
aiosqlite==0.21.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
bcrypt==4.0.1
cffi==2.0.0
click==8.1.8
cryptography==46.0.3
ecdsa==0.19.1
exceptiongroup==1.3.1
fastapi==0.123.5
greenlet==3.2.4
h11==0.16.0
idna==3.11
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
python-jose==3.5.0
python-multipart==0.0.20
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.44
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
